Save groups using /classification-api/v1/groups
This patch implements a command line utility implementing the groups Node
Classification endpoint. All groups currently defined in a Puppet Enterprise
2016.1 node classification service are sent to standard output.  The intent is
to save the groups into a pretty JSON format suitable for saving to Git version
control and also suitable for restore using the
[/v1/import-hierarchy](https://docs.puppet.com/pe/2016.1/nc_import-hierarchy.html)
endpoint.

This initial structure includes a command line utility, Application structure in
the Gem, and additional work related to spec testing, documentation, and code
coverage.

The goal is to make it as easy as possible to backup and restore node
classification groups.  This should be a matter of two steps on a monolithic
master:

    /opt/puppetlabs/puppet/bin/gem install ncio
    /opt/puppetlabs/puppet/bin/ncio backup > groups.$(date +%s).json

And to restore:

    /opt/puppetlabs/puppet/bin/ncio restore < groups.1467075714.json

TODO:

 * Update README

Completed work:

 * Application structure
 * Add rubocop and fix all warnings.
 * Configure RDoc to use YARD with Markdown.
 * Add simplecov to generate test coverage report
 * Test coverage
 * Logging facility