Add transform subcommand to restore on a different host
Without this patch it is not possible to restore a backup taken on another host.
This patch addresses the problem by transforming the backup data in the
following manner.  Node groups having class names that match the `--class-matcher`
command line option (A string passed to Regexp.new) are processed.  The value of
the `'rule'` key is recursively transformed, replacing all instances of the
specified hostnames.  Additionally, the class parameter values are
(non-recursively) transformed.

Transformation is fairly naive.  Literal string matches are replaced, no regular
expression or substring matching is performed on the hostname replacement.