Log backups and restores to syslog daemon facility
Without this patch there is no logging to syslog.  This is a problem because
backup and restore operations need to be monitored.  If not, people may lose
sleep wondering if the backup is successful or not.

This patch addresses the problem by logging to the syslog daemon facility by
default.  The result of the backup and the restore are logged at level "notice"
which is enabled by default in most logging configurations.  The goal is make
sure the result is written to the system log and tools like Splunk or Logstash
can easily keep an eye on the result.

Catch-all exception handlers are also added to the backup and restore task. This
allows log messages to be written in the event of any unforeseen errors.