Add ncio-replicate script intended for cron
This shell script is intended to be executed from cron on a periodic basis. The
goal is to keep a Standby PE Monolithic master in sync with an active Primary.
Pass the FQDN of the primary as ARG 1 and the FQDN of the secondary as ARG 2

In a DR situation when the secondary becomes active, block replication by
touching the lockfile.  This will prevent any changes made to the standby from
being clobbered as soon as the primary comes back online.

Prior to re-enabling replication after a DR situation, replicate back to the
primary by reversing the direction of this script.