Add backwards compatible default certificate behavior
Without this patch the default certificate used isn't backwards compatible with
previous versions of ncio.  This patch makes the behavior of how the default
certificate is handled backwards compatible with previous versions.

If the pe-internal-orchestrator certificate exists, it will be used as the
default.  If it does not exist, then the FQDN certificate will be used as the
default.