Make the fqdn cert the default
Without this patch the pe-internal-orchestrator certificate will take
precedence over the fqdn certificate if it exists.  This is a problem because a
PE 2016.4.3 server rebuilt from scratch, but having a SSL directory copied in
from an installation that had orchestrator enabled will fail to work.  It will
fail because the pe-internal-orchestrator certificate exists, and is trusted,
but it is not whitelisted.  Only the FQDN certificate is whitelisted on a
default PE 2016.4.3 install.

This patch makes ncio more robust by defaulting to use the fqdn certificate.